+++ dm-20151123/alibabacloud_dm20151123/models.py
@@ -94,9 +94,6 @@ class AddIpfilterResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -210,9 +207,6 @@ class ApproveReplyMailAddressResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -256,6 +250,8 @@ class BatchSendMailRequest(TeaModel):
         resource_owner_id: int = None,
         tag_name: str = None,
         template_name: str = None,
+        un_subscribe_filter_level: str = None,
+        un_subscribe_link_type: str = None,
     ):
         self.account_name = account_name
         self.address_type = address_type
@@ -268,6 +264,8 @@ class BatchSendMailRequest(TeaModel):
         self.resource_owner_id = resource_owner_id
         self.tag_name = tag_name
         self.template_name = template_name
+        self.un_subscribe_filter_level = un_subscribe_filter_level
+        self.un_subscribe_link_type = un_subscribe_link_type
 
     def validate(self):
         pass
@@ -300,6 +298,10 @@ class BatchSendMailRequest(TeaModel):
             result['TagName'] = self.tag_name
         if self.template_name is not None:
             result['TemplateName'] = self.template_name
+        if self.un_subscribe_filter_level is not None:
+            result['UnSubscribeFilterLevel'] = self.un_subscribe_filter_level
+        if self.un_subscribe_link_type is not None:
+            result['UnSubscribeLinkType'] = self.un_subscribe_link_type
         return result
 
     def from_map(self, m: dict = None):
@@ -326,6 +328,10 @@ class BatchSendMailRequest(TeaModel):
             self.tag_name = m.get('TagName')
         if m.get('TemplateName') is not None:
             self.template_name = m.get('TemplateName')
+        if m.get('UnSubscribeFilterLevel') is not None:
+            self.un_subscribe_filter_level = m.get('UnSubscribeFilterLevel')
+        if m.get('UnSubscribeLinkType') is not None:
+            self.un_subscribe_link_type = m.get('UnSubscribeLinkType')
         return self
 
 
@@ -374,9 +380,6 @@ class BatchSendMailResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -505,9 +508,6 @@ class CheckDomainResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -633,9 +633,6 @@ class CheckReplyToMailAddressResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -755,9 +752,6 @@ class CreateDomainResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -889,9 +883,6 @@ class CreateMailAddressResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1023,9 +1014,6 @@ class CreateReceiverResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1151,9 +1139,6 @@ class CreateTagResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1267,9 +1252,6 @@ class DeleteDomainResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1383,9 +1365,6 @@ class DeleteInvalidAddressResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1505,9 +1484,6 @@ class DeleteIpfilterByEdmIdResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1621,9 +1597,6 @@ class DeleteMailAddressResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1739,9 +1712,6 @@ class DeleteReceiverResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1861,9 +1831,6 @@ class DeleteReceiverDetailResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1979,9 +1946,6 @@ class DeleteTagResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -2185,9 +2149,6 @@ class DescAccountSummaryResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -2475,9 +2436,6 @@ class DescDomainResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -2591,9 +2549,6 @@ class GetIpProtectionResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -2801,9 +2756,6 @@ class GetIpfilterListResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -3113,9 +3065,6 @@ class GetTrackListResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -3437,9 +3386,6 @@ class GetTrackListByMailFromAndTagNameResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -3565,9 +3511,6 @@ class ModifyMailAddressResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -3705,9 +3648,6 @@ class ModifyPWByDomainResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -3836,9 +3776,6 @@ class ModifyTagResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -4118,9 +4055,6 @@ class QueryDomainByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -4352,9 +4286,6 @@ class QueryInvalidAddressResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -4640,9 +4571,6 @@ class QueryMailAddressByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -4904,9 +4832,6 @@ class QueryReceiverByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -5144,9 +5069,6 @@ class QueryReceiverDetailResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -5372,9 +5294,6 @@ class QueryTagByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -5642,9 +5561,6 @@ class QueryTaskByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -5846,9 +5762,6 @@ class SaveReceiverDetailResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -6004,9 +5917,6 @@ class SendTestByTemplateResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -6250,9 +6160,6 @@ class SenderStatisticsByTagNameAndBatchIDResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -6514,9 +6421,6 @@ class SenderStatisticsDetailByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -6564,6 +6468,8 @@ class SingleSendMailRequest(TeaModel):
         tag_name: str = None,
         text_body: str = None,
         to_address: str = None,
+        un_subscribe_filter_level: str = None,
+        un_subscribe_link_type: str = None,
     ):
         self.account_name = account_name
         self.address_type = address_type
@@ -6580,6 +6486,8 @@ class SingleSendMailRequest(TeaModel):
         self.tag_name = tag_name
         self.text_body = text_body
         self.to_address = to_address
+        self.un_subscribe_filter_level = un_subscribe_filter_level
+        self.un_subscribe_link_type = un_subscribe_link_type
 
     def validate(self):
         pass
@@ -6620,6 +6528,10 @@ class SingleSendMailRequest(TeaModel):
             result['TextBody'] = self.text_body
         if self.to_address is not None:
             result['ToAddress'] = self.to_address
+        if self.un_subscribe_filter_level is not None:
+            result['UnSubscribeFilterLevel'] = self.un_subscribe_filter_level
+        if self.un_subscribe_link_type is not None:
+            result['UnSubscribeLinkType'] = self.un_subscribe_link_type
         return result
 
     def from_map(self, m: dict = None):
@@ -6654,6 +6566,10 @@ class SingleSendMailRequest(TeaModel):
             self.text_body = m.get('TextBody')
         if m.get('ToAddress') is not None:
             self.to_address = m.get('ToAddress')
+        if m.get('UnSubscribeFilterLevel') is not None:
+            self.un_subscribe_filter_level = m.get('UnSubscribeFilterLevel')
+        if m.get('UnSubscribeLinkType') is not None:
+            self.un_subscribe_link_type = m.get('UnSubscribeLinkType')
         return self
 
 
@@ -6702,9 +6618,6 @@ class SingleSendMailResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -6818,9 +6731,6 @@ class UpdateIpProtectionResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()